requirements.txt
SpeechRecognition
pyttsx3
google-generativeai
requests
pyautogui
python-dotenv
pytest
pytest-mock
responses